requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
requests==2.31.0
pymongo==4.6.1 